// src/generated/definitions/APIGrantDetail.ts
/**
 * Do not edit this file it is auto-generated by io-utils / gen-api-models.
 * See https://github.com/pagopa/io-utils
 */
/* eslint-disable  */

import { OrganizationFiscalCode } from "./OrganizationFiscalCode.js";
import { ServiceId } from "./ServiceId.js";
import { FiscalCode } from "./FiscalCode.js";
import * as t from "io-ts";
import { NonEmptyString } from "@pagopa/ts-commons/lib/strings.js";
import { UTCISODateFromString } from "@pagopa/ts-commons/lib/dates.js";

// required attributes
const APIGrantDetailR = t.interface({
  id: NonEmptyString,

  issuedAt: UTCISODateFromString,

  expireAt: UTCISODateFromString,

  organizationId: OrganizationFiscalCode,

  serviceId: ServiceId,

  identityId: FiscalCode,

  scope: t.readonlyArray(t.string, "array of string"),
});

// optional attributes
const APIGrantDetailO = t.partial({});

export const APIGrantDetail = t.intersection(
  [APIGrantDetailR, APIGrantDetailO],
  "APIGrantDetail"
);

export type APIGrantDetail = t.TypeOf<typeof APIGrantDetail>;